src/app/forms.py
from flask_wtf import FlaskForm
from wtforms import StringField, PasswordField, SubmitField
from wtforms.validators import DataRequired, Email, Length

class LoginForm(FlaskForm):
    email = StringField("Email", validators=[DataRequired(), Email()])
    password = PasswordField("Password", validators=[DataRequired()])
    submit = SubmitField("Login")

class RegisterForm(FlaskForm):
    name = StringField("Name", validators=[DataRequired(), Length(min=2, max=256)])
    email = StringField("Email", validators=[DataRequired(), Email(), Length(min=6)])
    password = PasswordField("Password", validators=[DataRequired(), Length(min=6)])
    register = SubmitField("Register")