src/requirements.txt
flask==3.1.3
python-dotenv==1.2.2
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-bcrypt==1.0.1
flask-wtf==1.2.2
email-validator==2.3.0
flask-migrate==4.1.0
psycopg2-binary==2.9.11
pytest==9.0.2
pytest-flask==1.3.0